python_rag_service/requirements.txt
google-generativeai
Flask
chromadb
requests
beautifulsoup4
lxml
APScheduler
pandas